pipeline/requirements.txt
backports-abc==0.4
beautifulsoup4==4.5.1
biocommons.seqrepo>=0.4.4
ga4gh.vr
biopython==1.69
bioutils==0.3.3
bx-python==0.7.3
certifi>=2017.4.17
configparser==3.5.0
CrossMap==0.2.6
Cython==0.24.1
docutils==0.12
hgvs==1.3.0
intervaltree==3.0.2
itertoolz==0.5
lxml==4.2.5
lockfile==0.12.2
luigi==2.8.2
nose==1.3.7
numpy==1.14.*
Parsley==1.3
pip==9.0.1
pygr==0.8.2
PyJWT==1.4.1
pysam==0.9.1.4
python-daemon==2.1.1
PyVCF==0.6.8
recordtype==1.1
requests==2.20.0
retrying==1.3.3
setuptools==25.1.4
singledispatch==3.4.0.3
six==1.10.0
pandas==0.23.4
tornado==4.4.1
toolz==0.9.0
wheel==0.29.0
click==6.7
